Retro std::size_t getUnitsSize() to int

+++ src/map/cMap.cpp
@@ -572,7 +572,7 @@ void cMap::draw_units()
     //// @Mira fix trasnparency set_trans_blender(0, 0, 0, 160);
 
     // draw all worms first
-    for (size_t i = 0; i < m_objects->getUnitsSize(); i++) {
+    for (int i = 0; i < m_objects->getUnitsSize(); i++) {
         cUnit *pUnit = m_objects->getUnit(i);
         if (!pUnit || !pUnit->isValid()) continue;
 
@@ -592,7 +592,7 @@ void cMap::draw_units()
     }
 
     // then: draw infantry units
-    for (size_t i = 0; i < m_objects->getUnitsSize(); i++) {
+    for (int i = 0; i < m_objects->getUnitsSize(); i++) {
         cUnit *pUnit = m_objects->getUnit(i);
         if (!pUnit || !pUnit->isValid()) continue;
 
@@ -608,7 +608,7 @@ void cMap::draw_units()
     }
 
     // then: draw ground units
-    for (size_t i = 0; i < m_objects->getUnitsSize(); i++) {
+    for (int i = 0; i < m_objects->getUnitsSize(); i++) {
         cUnit *pUnit = m_objects->getUnit(i);
         if (!pUnit || !pUnit->isValid()) continue;
 
@@ -640,7 +640,7 @@ void cMap::draw_units_2nd()
     auto *mapViewport = m_interface->getMapViewport();
 
     // draw health of units
-    for (size_t i = 0; i < m_objects->getUnitsSize(); i++) {
+    for (int i = 0; i < m_objects->getUnitsSize(); i++) {
         cUnit *pUnit = m_objects->getUnit(i);
         if (!pUnit || !pUnit->isValid()) continue;
         if (!pUnit->rendering.bHovered && !pUnit->isSelected()) continue;
@@ -656,7 +656,7 @@ void cMap::draw_units_2nd()
     }
 
     // draw airborn units
-    for (size_t i = 0; i < m_objects->getUnitsSize(); i++) {
+    for (int i = 0; i < m_objects->getUnitsSize(); i++) {
         cUnit *pUnit = m_objects->getUnit(i);
         if (!pUnit || !pUnit->isValid()) continue;
         if (!pUnit->isAirbornUnit()) continue;
